refactor: (#169) shop 관련 로직 같은 slice로 통합
