--- src/features/cart/index.ts
@@ -1,2 +1,0 @@
-export { useCart } from './model/useCart';
-export type { CartItem } from './model/useCart';
