--- src/features/shop-purchase/index.ts
@@ -1,2 +1,0 @@
-export { ShopPurchaseButton } from './ui/ShopPurchaseButton';
-export { purchaseQueries } from './queries/purchaseQueries';
