--- src/features/shop-purchase/model/types.ts
@@ -1,1 +1,0 @@
-export type PurchaseStatus = 'IDLE' | 'SUCCESS' | 'FAIL_BALANCE' | 'FAIL_LEVEL';
