+++ src/features/shop/model/useCart.ts
@@ -3,7 +3,7 @@ import type { Product } from '@/entities/product';
 import { useSound } from '@/entities/sound';
 import { SoundManager } from '@/shared/api/sound/manager';
 import { SOUND_ASSETS } from '@/shared/api/sound/assets';
-import { SHOP_CATEGORIES } from '@/features/shop';
+import { SHOP_CATEGORIES } from '../constants/shop.constants';
 
 export interface CartItem extends Product {
   quantity: number;
