+++ src/features/shop/queries/shopQueries.ts
@@ -1,8 +1,9 @@
-import { queryOptions } from '@tanstack/react-query';
+import { mutationOptions, queryOptions } from '@tanstack/react-query';
 
-import { getConsumables } from '@/features/shop/api/getConsumables';
-import { getCosmetics } from '@/features/shop/api/getCosmetics';
-import { getUserInventory } from '@/features/shop/api/getUserInventory';
+import { getConsumables } from '../api/getConsumables';
+import { getCosmetics } from '../api/getCosmetics';
+import { getUserInventory } from '../api/getUserInventory';
+import { postPurchase } from '../api/postPurchase';
 
 export const shopQueries = {
   all: () => ['shop'] as const,
@@ -23,5 +24,11 @@ export const shopQueries = {
     queryOptions({
       queryKey: [...shopQueries.all(), 'inventory'] as const,
       queryFn: getUserInventory,
+    }),
+
+  purchase: () =>
+    mutationOptions({
+      mutationKey: [...shopQueries.all(), 'purchase'] as const,
+      mutationFn: postPurchase,
     }),
 };
