+++ src/features/shop/ui/PurchaseStatusContent.tsx
