+++ src/features/shop/ui/ReceiptItemCarousel.tsx
